video feature backend end points

+++ back-end/server/messageRoutes.js
@@ -12,14 +12,8 @@ router.get("/getSendDriftBottles/:userId", async (req, res) => {
 	res.send(sendDriftBottles)
 })
 
-router.post ("/addDriftBottlesImage", (req, res) => {
-    const fullServerUrl = req.protocol + '://' + req.get('host');
-    const form = new multiparty.Form();
-    form.parse(req, async (err,fields, files) => {
-        const imageFileName = fields.imageFileName[0];
-        const userId = fields.userId[0];
-        const bottleId = fields.bottleId[0];
-        let hashImageName = null;
+
+const getDateString = () => {
         const date_ob = new Date();
         let date = ("0" + date_ob.getDate()).slice(-2);
 
@@ -37,8 +31,25 @@ router.post ("/addDriftBottlesImage", (req, res) => {
 
         // current seconds
         let seconds = date_ob.getSeconds();
+    
+        return  year + month + date + hours +  minutes +  seconds;
+}
+
+const getHashFileName = (userId, fileName) => {
+    return userId + "-" + getDateString() + Math.floor(Math.random() * 100) + 1 + fileName;
+}
+
+router.post ("/addDriftBottlesImages", (req, res) => {
+    const fullServerUrl = req.protocol + '://' + req.get('host');
+    const form = new multiparty.Form();
+    form.parse(req, async (err,fields, files) => {
+        const imageFileName = fields.imageFileName[0];
+        const userId = fields.userId[0];
+        const bottleId = fields.bottleId[0];
+        let hashImageName = null;
         
-        hashImageName = userId + "-" + year + month + date + hours +  minutes +  seconds + Math.floor(Math.random() * 100) + 1 + imageFileName;
+        
+        hashImageName = getHashFileName(userId, imageFileName);
         //The file itself
         const uploadImageFile = files.imageFile[0];
         const path_temp = uploadImageFile.path;
@@ -55,7 +66,41 @@ router.post ("/addDriftBottlesImage", (req, res) => {
             else console.log("Update image file name and url successfully");
         });
         
-        res.json("AddDriftBottlesImage SUCCESS")
+        res.json("AddDriftBottlesImages SUCCESS")
+        
+    
+    });
+
+})
+
+router.post ("/addDriftBottlesVideos", (req, res) => {
+    const fullServerUrl = req.protocol + '://' + req.get('host');
+    const form = new multiparty.Form();
+    form.parse(req, async (err,fields, files) => {
+        const videoFileName = fields.videoFileName[0];
+        const userId = fields.userId[0];
+        const bottleId = fields.bottleId[0];
+        let hashVideoName = null;
+        
+        
+        hashVideoName = getHashFileName(userId, videoFileName);
+        //The file itself
+        const uploadVideoFile = files.videoFile[0];
+        const path_temp = uploadVideoFile.path;
+        //Move uploaded file with unique name to server public/files folder
+        await fs.rename(path_temp,`${__dirname}/public/driftBottleVideos/${hashVideoName}`, (err) => {
+            if (err) console.error(err);
+        });
+        
+        const query = {_id: bottleId},
+        update = {videoFileName: hashVideoName, videoUrl: fullServerUrl+"/messageApi/getDriftBottleVideos/"+hashVideoName},
+        options = { upsert: true, new: true, setDefaultsOnInsert: true };
+        await SendDriftBottles.findOneAndUpdate(query, update, options, (error, result) => {
+            if (error) console.log("Failed to update video file name and url. Error: " + error.stack);
+            else console.log("Update video name and url successfully");
+        });
+        
+        res.json("AddDriftBottlesVideos SUCCESS")
         
     
     });
@@ -99,7 +144,32 @@ router.get('/getDriftBottleImages/:imageName', function (req, res, next) {
         if (err) {
             next(err);
         } else {
-            console.log('Sent: ', fileName);
+            console.log('Send Drift Bottle Image: ', fileName);
+        }
+    })
+
+});
+
+
+router.get('/getDriftBottleVideos/:videoName', function (req, res, next) {
+
+    const options = {
+        root: __dirname + '/public/driftBottleVideos/',
+        dotfiles: 'deny',
+        headers: {
+            'x-timestamp': Date.now(),
+            'x-sent': true
+        }
+    };
+    const fileName = req.params.videoName;
+
+
+       //Send request file name
+    res.sendFile(fileName, options, (err) => {
+        if (err) {
+            next(err);
+        } else {
+            console.log('Send Drift Bottle Video: ', fileName);
         }
     })
 
@@ -108,12 +178,24 @@ router.get('/getDriftBottleImages/:imageName', function (req, res, next) {
 router.delete("/deleteSendDriftBottles/:id", async (req, res) => {
 	try {
         
-        const bottleToDelete = await SendDriftBottles.find({_id:req.params.id}, 'imageFileName');
-        if (bottleToDelete && bottleToDelete[0].imageFileName) {
-           let filePath = `${__dirname}/public/driftBottleImages/${bottleToDelete[0].imageFileName}`;
-        await fs.unlink(filePath, (error) => {
-            if(error) console.error(error);
-        }) 
+        // Delete image if any
+        const bottleToDeleteImage = await SendDriftBottles.find({_id:req.params.id}, 'imageFileName');
+        
+        if (bottleToDeleteImage && bottleToDeleteImage[0].imageFileName) {
+           let filePath = `${__dirname}/public/driftBottleImages/${bottleToDeleteImage[0].imageFileName}`;
+            await fs.unlink(filePath, (error) => {
+                if(error) console.error(error);
+            }) 
+        }
+        
+        // Delete video if any
+        const bottleToDeleteVideo = await SendDriftBottles.find({_id:req.params.id}, 'videoFileName');
+        
+        if (bottleToDeleteVideo && bottleToDeleteVideo[0].videoFileName) {
+           let filePath = `${__dirname}/public/driftBottleVideos/${bottleToDeleteVideo[0].videoFileName}`;
+            await fs.unlink(filePath, (error) => {
+                if(error) console.error(error);
+            }) 
         }
         
 		await SendDriftBottles.deleteOne({ _id: req.params.id })
